# api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CS2KZ API",
    "description": "Source Code available on [GitHub](https://github.com/KZGlobalTeam/cs2kz-api).",
    "contact": {
      "name": "AlphaKeks",
      "email": "alphakeks@dawn.sh"
    },
    "license": {
      "name": "Licensed under the GPLv3",
      "url": "https://www.gnu.org/licenses/gpl-3.0"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/bans": {
      "get": {
        "tags": [
          "Bans"
        ],
        "summary": "Fetch bans.",
        "operationId": "get",
        "parameters": [
          {
            "name": "player",
            "in": "query",
            "description": "Filter by player.",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/PlayerIdentifier"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "server",
            "in": "query",
            "description": "Filter by server.",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ServerIdentifier"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "reason",
            "in": "query",
            "description": "Filter by ban reason.",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/BanReason"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "unbanned",
            "in": "query",
            "description": "Filter by bans that have already been reverted.",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "banned_by",
            "in": "query",
            "description": "Filter by admins who issued bans.",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/PlayerIdentifier"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "unbanned_by",
            "in": "query",
            "description": "Filter by admins who reverted bans.",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/PlayerIdentifier"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "created_after",
            "in": "query",
            "description": "Only include bans submitted after this date.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            }
          },
          {
            "name": "created_before",
            "in": "query",
            "description": "Only include bans submitted before this date.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "total",
                    "results"
                  ],
                  "properties": {
                    "total": {
                      "type": "integer",
                      "format": "uint64",
                      "minimum": 0
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "description": "A player ban.",
                        "required": [
                          "id",
                          "player",
                          "reason",
                          "created_on"
                        ],
                        "properties": {
                          "id": {
                            "$ref": "#/components/schemas/BanID"
                          },
                          "player": {
                            "$ref": "#/components/schemas/Player"
                          },
                          "server": {
                            "allOf": [
                              {
                                "$ref": "#/components/schemas/ServerInfo"
                              }
                            ],
                            "nullable": true
                          },
                          "reason": {
                            "$ref": "#/components/schemas/BanReason"
                          },
                          "admin": {
                            "allOf": [
                              {
                                "$ref": "#/components/schemas/Player"
                              }
                            ],
                            "nullable": true
                          },
                          "created_on": {
                            "type": "string",
                            "format": "date-time",
                            "description": "When this ban was submitted."
                          },
                          "expires_on": {
                            "type": "string",
                            "format": "date-time",
                            "description": "When this ban will expire.",
                            "nullable": true
                          },
                          "unban": {
                            "allOf": [
                              {
                                "$ref": "#/components/schemas/Unban"
                              }
                            ],
                            "nullable": true
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Bans"
        ],
        "summary": "Create a new ban.",
        "operationId": "post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewBan"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Response body for submitting a new ban.",
                  "required": [
                    "ban_id"
                  ],
                  "properties": {
                    "ban_id": {
                      "$ref": "#/components/schemas/BanID"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "422": {
            "description": ""
          }
        },
        "security": [
          {
            "Browser Session": [
              "bans"
            ]
          }
        ]
      }
    },
...
    "/plugin/versions": {
      "get": {
        "tags": [
          "CS2KZ Plugin"
        ],
        "summary": "Fetch CS2KZ plugin versions.",
        "operationId": "get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "total",
                    "results"
                  ],
                  "properties": {
                    "total": {
                      "type": "integer",
                      "format": "uint64",
                      "minimum": 0
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "description": "A CS2KZ plugin version.",
                        "required": [
                          "id",
                          "semver",
                          "git_revision",
                          "created_on"
                        ],
                        "properties": {
                          "id": {
                            "$ref": "#/components/schemas/PluginVersionID"
                          },
                          "semver": {
                            "type": "string",
                            "description": "The version as a string."
                          },
                          "git_revision": {
                            "type": "string",
                            "description": "The git revision associated with this release."
                          },
                          "created_on": {
                            "type": "string",
                            "format": "date-time",
                            "description": "When this version was submitted."
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "CS2KZ Plugin"
        ],
        "summary": "Submit a new CS2KZ plugin version.",
        "description": "This endpoint is intended to be used by GitHub Actions.",
        "operationId": "post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPluginVersion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Response body for submitting a new plugin version.",
                  "required": [
                    "plugin_version_id"
                  ],
                  "properties": {
                    "plugin_version_id": {
                      "$ref": "#/components/schemas/PluginVersionID"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "422": {
            "description": ""
          }
        },
        "security": [
          {
            "API Key": [
              "plugin_versions"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "BanID": {
        "type": "integer",
        "format": "uint64",
        "minimum": 0
      },
      "BanReason": {
        "type": "string",
        "description": "Ban reasons.",
        "enum": [
          "auto_strafe",
          "auto_bhop"
        ]
      },
...
      "Limit": {
        "type": "number",
        "description": "limits the amount of returned values",
        "minimum": 0
      },
...
      "NewBan": {
        "type": "object",
        "description": "Request payload for submitting a new ban.",
        "required": [
          "player_id",
          "reason"
        ],
        "properties": {
          "player_id": {
            "$ref": "#/components/schemas/SteamID"
          },
          "player_ip": {
            "type": "string",
            "description": "The IP address of the player who should be banned.",
            "nullable": true
          },
          "reason": {
            "$ref": "#/components/schemas/BanReason"
          }
        }
      },
...
      "NewPluginVersion": {
        "type": "object",
        "description": "Request payload for submitting a new plugin version.",
        "required": [
          "semver",
          "git_revision"
        ],
        "properties": {
          "semver": {
            "type": "string",
            "description": "The version as a string."
          },
          "git_revision": {
            "type": "string",
            "description": "The git revision associated with this release."
          }
        }
      },
...
      "Offset": {
        "type": "number",
        "description": "used for pagination",
        "default": 0,
        "maximum": 1.7976931348623157e308,
        "minimum": -1.7976931348623157e308
      },
      "Player": {
        "type": "object",
        "description": "Basic information about a KZ player.",
        "required": [
          "name",
          "steam_id"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The player's name."
          },
          "steam_id": {
            "$ref": "#/components/schemas/SteamID"
          }
        }
      },
      "PlayerIdentifier": {
        "oneOf": [
          {
            "oneOf": [
              {
                "type": "string",
                "title": "Steam ID",
                "example": "STEAM_1:1:161178172"
              },
              {
                "type": "string",
                "title": "Steam ID3",
                "example": "U:1:322356345"
              },
              {
                "type": "integer",
                "title": "Steam ID32",
                "example": 322356345
              },
              {
                "type": "integer",
                "title": "Steam ID64",
                "example": 76561198282622073
              }
            ],
            "description": "a player's SteamID",
            "example": "STEAM_1:1:161178172"
          },
          {
            "type": "string",
            "title": "Name"
          }
        ],
        "description": "A SteamID or name"
      },
...
      "PluginVersionID": {
        "type": "integer",
        "format": "uint16",
        "minimum": 0
      },
...
      "ServerID": {
        "type": "integer",
        "format": "uint16",
        "minimum": 0
      },
      "ServerIdentifier": {
        "oneOf": [
          {
            "type": "integer",
            "title": "ID"
          },
          {
            "type": "string",
            "title": "Name"
          }
        ],
        "description": "A server ID or name"
      },
      "ServerInfo": {
        "type": "object",
        "description": "Information about a KZ server.",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ServerID"
          },
          "name": {
            "type": "string",
            "description": "The server's name."
          }
        }
      },
...
      "SteamID": {
        "oneOf": [
          {
            "type": "string",
            "title": "Steam ID",
            "example": "STEAM_1:1:161178172"
          },
          {
            "type": "string",
            "title": "Steam ID3",
            "example": "U:1:322356345"
          },
          {
            "type": "integer",
            "title": "Steam ID32",
            "example": 322356345
          },
          {
            "type": "integer",
            "title": "Steam ID64",
            "example": 76561198282622073
          }
        ],
        "description": "a player's SteamID",
        "example": "STEAM_1:1:161178172"
      },
...
      "Unban": {
        "type": "object",
        "description": "Reversion of a `Ban`.",
        "required": [
          "id",
          "reason",
          "created_on"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UnbanID"
          },
          "reason": {
            "type": "string",
            "description": "The reason for the unban."
          },
          "admin": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Player"
              }
            ],
            "nullable": true
          },
          "created_on": {
            "type": "string",
            "format": "date-time",
            "description": "When this ban was reverted."
          }
        }
      },
      "UnbanID": {
        "type": "integer",
        "format": "uint64",
        "minimum": 0
      },
...
    },
    "securitySchemes": {
      "API Key": {
        "type": "http",
        "scheme": "bearer"
      },
      "Browser Session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kz-auth"
      },
      "CS2 Server": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}